feat: adds code from Giraffe
Since this code 'stands-on-its-own-two-feet' it has been pulled out of
Giraffe and converted into its own project.

This will slightly reduce the size of the Giraffe codebase since another
templating engine may be used.

Since all of the namespacing is maintained, it can still be dropped into
a giraffe project with zero code modifications required (just reference
the code in the dependencies)